back_end/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.2
django-timezone-field==5.0
djangorestframework==3.14.0
flower==1.2.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
humanize==4.6.0
install==1.3.5
kombu==5.3.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
validate-docbr==1.10.0
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
